+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Version 0.54.3
 **Bugfixes**
 - Support device class `APPLE_SILICON_MAC` on App Store Connect API [`Device`](https://developer.apple.com/documentation/appstoreconnectapi/device) responses. [PR #437](https://github.com/codemagic-ci-cd/cli-tools/pull/437)
 
+**Improvements**
+- Fail action `app-store-connect fetch-signing-files` early with descriptive error message if bundle ID identifier is not given. [PR #438](https://github.com/codemagic-ci-cd/cli-tools/pull/438)
+
 Version 0.54.2
 -------------